# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Swagger API documentation",
        "contact": {
            "email": "igornosatov@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/library": {
            "get": {
                "tags": [
                    "Book API"
                ],
                "summary": "Display list of the books",
                "operationId": "booksAll",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by book title or author name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Order by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort by",
                        "in": "query",
                        "description": "Sort by asc/desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Show books by genre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_paginate",
                        "in": "query",
                        "description": "Paginate the data",
                        "schema": {
                            "type": "string",
                            "default": "true",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/library/store": {
            "post": {
                "tags": [
                    "Book API"
                ],
                "summary": "Book create",
                "description": "API for creating Book",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "book title",
                                        "type": "string"
                                    },
                                    "author": {
                                        "description": "book author",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "book description",
                                        "type": "string"
                                    },
                                    "genre": {
                                        "description": "genre",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success creating data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/library/edit/{id}": {
            "get": {
                "tags": [
                    "Book API"
                ],
                "summary": "Edit book",
                "operationId": "editBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success update data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/library/{id}/update": {
            "patch": {
                "tags": [
                    "Book API"
                ],
                "summary": "Update book",
                "description": "Update Book",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "string",
                                        "format": "binary",
                                        "example": "1.jpeg"
                                    },
                                    "title": {
                                        "description": "book title",
                                        "type": "string",
                                        "example": "book title"
                                    },
                                    "author": {
                                        "description": "book author",
                                        "type": "string",
                                        "example": "book author"
                                    },
                                    "description": {
                                        "description": "book description",
                                        "type": "string",
                                        "example": "book description"
                                    },
                                    "genre": {
                                        "description": "genre",
                                        "type": "string",
                                        "example": "novel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success creating data"
                    },
                    "400": {
                        "description": "Invalid request data"
                    }
                }
            }
        },
        "/library/delete/{id}": {
            "delete": {
                "tags": [
                    "Book API"
                ],
                "summary": "Delete book",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success deleting data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Books"
        }
    ]
}